// Source/SelfService/Web/apis/integrations/generated/models/RestApiServiceStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aigonix.Bridge.M3
 * Bridge API - made for Aigonix Studio
 *
 * The version of the OpenAPI document: 0.0.1.351
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const RestApiServiceStatus = {
    Off: 'Off',
    Deploying: 'Deploying',
    Active: 'Active',
    Unhealthy: 'Unhealthy',
    Terminating: 'Terminating'
} as const;
export type RestApiServiceStatus = typeof RestApiServiceStatus[keyof typeof RestApiServiceStatus];


export function RestApiServiceStatusFromJSON(json: any): RestApiServiceStatus {
    return RestApiServiceStatusFromJSONTyped(json, false);
}

export function RestApiServiceStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): RestApiServiceStatus {
    return json as RestApiServiceStatus;
}

export function RestApiServiceStatusToJSON(value?: RestApiServiceStatus | null): any {
    return value as any;
}